api/requirements.txt
together
sentence-transformers
numpy
fastapi
uvicorn
python-dotenv
requests